Specials: Show preview image on Special:Upload left hand for rtl lang
Visit Special:Upload?uselang=rtl and choose a file to upload that is
supported by the file preview javascript (not to big).
The preview image is on the right hand and moves all text to the left.
Fix this by using own css selector and allow flip,
the used mw-halign-right is for content and is not allowed to flip.
Also use the css selector when showing a stashed file preview.
UploadForm is used by extensions, keep signature

Bug: T410766
Change-Id: I6620da56a886599fa15d77ff8d1fb522b6f681f8