Don't enable parser options that are set to false
This came up while working on I1aef3fa. As of now the values in an
[ ParserOutputAccess::OPT_NO_UPDATE_CACHE => false ] array are
ignored and the options always enabled, no matter what the array
value says. I feel like this is confusing, as well as a missed
oportunity.

I'm also proposing a little bit of code cleanup in the
normalizeOptions() function. There is no point in setting values that
have no corresponding numeric OPT_… constant. There is no option 0
and (currently) no option larger than 128.

This code was originally written in Ibf8c1bd just a few months ago.

Bug: T387478
Change-Id: If6a44c069b45bcb54b6e5cc181cabee8da7d891e