ConverterRule: Separate construction from ConverterRule::parse()
This allows a future patch to add an alternative parse() method
suitable for MediaWiki DOM Spec elements.

Changing the method signature on ConverterRule::parse() is not a
breaking change as the ConverterRule class is not exposed outside
of mediawiki-core and is not @newable.

As a sanity check, ConverterRule is not used outside of mediawiki-core:

https://codesearch.wmcloud.org/deployed/?q=ConverterRule

and ConverterRule::parse() is not used outside of LanguageConverter:

https://huma.wmcloud.org/callers?class_name=MediaWiki%5CLanguage%5CConverterRule&method_name=parse

Change-Id: I7b00ecf2aee109fae01c6d6489c34c8b0115f813