logging: Add isLegacy to the LogEntry interface
The method was previously defined in LogEntryBase, which is the only
known implementor of LogEntry, and therefore all known implementations
had it. However, it was not defined on the interface. Therefore, typing
objects against the interface would not allow calling `isLegacy`.

Add the method to the interface, with a return type declaration as is
more standard in modern code, adjusting inheritors to avoid incompatible
signatures. None of these classes is stable to extend (and LogEntry is
not stable to implement), so these changes are not covered by the stable
interface policy.

In LogFormatter, type the `$entry` property as LogEntry, as that's what
the constructor types it against.

Change-Id: Ic82cc18e64245e5f5cfd45b84d75f0e4bbf55eb3