Add post-processing cache, disabled by default
This patch adds a post-processing cache that aims at storing some common
outputs of the post-processing pipeline.
It is disabled by default; this patch should have no effect on the
current cache situation and can be enabled via the UsePostprocCache
variable as well as the postproc-pcache configuration options.

To enable this feature, UsePostProcCache must be set to true, and
ParserCacheFilterConfig['postproc-pcache]['default']['minCpuTime'] must
be set to 0, as well as
ParserCacheFilterConfig['postproc-parsoid-pcache]['default']['minCpuTime'] .

Bug: T348255
Change-Id: Icf7172be6d718362192ca8b7288ef8ce4d4af6b9