HTMLFormFieldCloner: Fix multiple bugs related to conditional states
I don't like only fixing the symptom of PHP warnings, so I dug into
it and found the root cause. The default value is used instead of
null, like in the `HTMLFormField::getNearestFieldValue()` method.

After writing the test case, I realised the value of fields in the
cloner that are supposed to be disabled are not reset to the default,
so I also fixed it.

Also:
* Avoided (ab)using field data for passing internal state of 'nonjs',
  checked with codesearch and nobody outside use it.
* Updated the coverage markup for HTMLFormFieldTest.

Bug: T391882
Change-Id: Ia97b04387ebb76814f4483600094d2fe64995382