Pager: Clamp date overflow to avoid invalid offset timestamp
In ReverseChronologicalPager::getOffsetDate, the upper-bound timestamp
is computed by advancing the given year/month/day to the start of the
next day, month, or year as appropriate. However, this allows inputs
such as 9999-12-31 to advance to year 10000, which cannot be handled
by the Timestamp library.

This patch clamps the result in such cases to prevent the overflow.

Bug: T402541
Change-Id: I810a7589bdc8baeffa8f5f4d87aed6d8dc68380a