Add LookupLanguageSelector component
The LanguageSelector component uses a headless component approach
where the primary logic for a language selector is encapsulated
within a renderless component. The headless component can then be
used to build different interfaces for a language selector.

In this patch, a new Codex Lookup based language selector is added.

Limited customizability of the langauge selector is possible. Further
customizability can be introduced as needed.

Two modules are added in Resources.php:
* mediawiki.languageselector: The language selector
* mediawiki.codex.deps.languageselector: The dependencies of the
  language selector. Instead of adding the entire @wikimedia/codex
  as a dependency, specific Codex modules are added as dependency.

Tests are added, generated via LLM with manual fixes applied.

Bug: T410267
Change-Id: Ic9088672c4b1479bf051eb0f4ffda463534e8a81