REST: add "site" module, contining "sitemap" endpoints
Two "sitemap" endpoints exist in core under these paths:
 * site/v1/sitemap/{indexId}
 * site/v1/sitemap/{indexId}/page/{fileId}

By convention, these should really be in a module named "site".
Introduce a "site" module to contain these endpoints (as well as
any other future endpoints we may decide to put there).

This change intentionally does not remove the existing route
definitions from coreRoutes.json. This is because we need the new
module to be in the production codebase before we make the config
change to activate it. The existing routes will be removed in a
future change, after the new "site" module code is fully deployed.
Having an inactive module (or even an active one) with routes that
duplicate coreRoutes.json causes no errors.

Bug: T409516
Change-Id: Ib2371e256883f5bdb4058c41b3fdf4e11d1851fd