mediawiki.util: Add adjustThumbWidthForSteps for step sizing in JS
Added mw.util.adjustThumbWidthForSteps() which follows the same
logic as File::adjustThumbWidthForSteps in the internal thumbnail
logic on the PHP side.

This can be easily exposed to extensions like Popups that handle
thumbnail URL generation manually on the client side to make them
round up to the wgThumbnailSteps stops.

Does not follow wgThumbnailStepsRatio except for 0 and 1; non-
integral ratios will be treated the same as 1.

Popups in particular was making 320px, 480px, and 640px requests
for many images, as well as possibly other sizes for different
aspect ratios; this will bump them up to 330px, 500px, and 960px
respectively but be more likely to be pre-rendered.

A qunit test case is added duplicating the PHP-side test logic,
and both implementation and tests cross-reference each other in
comments to prevent future desyncing.

Possible follow-up work:

* add a common 'fit this box' thumbnail sizer to replace the
  parseImageUrl/resizeUrl code that's already there, and wrap
  this into it.

* update various extensions to use this code

Bug: T411013
Change-Id: I9a9a8b888a69505cb5d1babd7e1606761f68456c