user: Handle replag in UserOptionsManager::saveOptions()
UserOptionsManager::saveOptions() takes a UserIdentity, but
then uses it for a CAS check so callers had to jump through
hoops to reduce the likelyhood of conflicts. The value of a
CAS check for user options changes is minimal, but we do need
to bump user_touched since many user options influence how
pages are rendered so caches need to be invalidated.

Simplify things by making sure on the UserOptionsManager side
that the UserIdentity is converted to a primary record.

Core callers are updated to avoid unnecessary primary lookups;
this results in the user object passed to some hooks
(ApiOptions, GetPreferences, LoadUserOptions) being a replica
instance rather than a primary one.

Bug: T405231
Change-Id: I3c0d3d011569204ce08a14b173b8e9f5f0bc6de7