Specials: Fix handling of option for forcing global detach
In SpecialRenameUser, when warnings are shown, the user can choose
to ignore them by checking a confirmation checkbox attached to the
warning response; this simply works by not calling the function that
generates the warnings in the first place.

However, for global account detachment warning (which is added
by CentralAuth), this simple skip is not working because RenameUser
will redo check for global detaching. To support suppressing the
warning, 'forceGlobalDetach' option must be passed when constructing
the RenameUser instance.

Now, when there are warnings to be shown, this patch:
    1. Checks if there's a warning for global account detach
    2. If no, do nothing
    3. If yes, add a hidden `forceglobaldetach` field, so that we
       can pass the value to the subsequent request
    4. If we need to suppress the warnings, pass the value
       in 'forceGlobalDetach' option when instantiating RenameUser

Bug: T407242
Change-Id: I4087b51902232d47c8e3b02b517aae1fde529b59