Watchlist labels: Add error when saving a duplicate name
Add a filter and more validation for the name field in the label
editing form.

Also change WatchlistLabelStore::save() to return a Status. In
most cases for the duplicate case this won't be used, because
we're querying for the dupe before trying to run the UPDATE query,
but if two are saved at nearly the same time it will.

Bug: T411028
Change-Id: I06888c34eba519b1d357af38da760cf351f8ad02