Api: Avoid UploadStashBadPathException after upload error
A file upload can be an identical file re-upload to the same file,
the error "fileexists-no-change" is detected near to
the end and the internal php file is already consumed/purged.
In this case there is nothing to stash.
But trying to stash gives an extra "stasherrors" key in the result with
code "uploadstash-exception" and the extra param
"UploadStashBadPathException". This can hide the real error for the
client logging and should be avoided.

Move the decision from ApiUpload to UploadBase to allow different checks
for each upload type. This case is only relevant for the file upload.

For the test case allow to assert also the api data of the IApiMessage

Bug: T270866
Change-Id: Iaefc449c38454ddb5d2ca5a797e79f61fcfca4ff