SECURITY: Match user lang callback during message expansion
CVE-2026-34091

During expansion of a message parameter that is itself another message,
userLangCallback is not copied from the parent language. This leads to an
accidental leak of the main context's language (the user, in most cases)
if the user happens to use a language different from the wiki's content
language.

This ensures that the userLangCallback of the parent Message extracting a
MessageValue parameter will be used to get the message for that
MessageValue.

Bug: T411305
Change-Id: I6643086b6a1b0118222c41d66a5c2716235ef11c