user: Do not query DB in getInstanceFromPrimary() for primary user
If the User object is already from the primary DB, avoid a pointless
lookup in User::getInstanceFromPrimary().

This is a change to the contract since the method was documented
as always returning a new instance, so it's technically a breaking
change (getInstanceFromPrimary() was introduced in 1.45 which is
still in beta, but getInstanceForUpdate() is also affected) but
there's no reason for callers to care about that.

Change-Id: Ib4543f025c885d2180465c4287bf83a24b96b1d6