REST: remove sitemap endpoints from flat routes in coreRoutes.json
The "sitemap" endpoints are being moved from coreRoutes.json to
the new "site" module. The site.v1 module is already active on
all wikis and serving the routes. This change removes the old,
now unused routes from coreRoutes.json.

This will cause the endpoints to disappear from the REST Sandbox.
The site.v1 module is not visible in the REST Sandbox. This is all
intentional. See the related phab task for details.

Bug: T409516
Change-Id: I5064cedd629eb9042d5c05bc7bccbbf9f5bd5d1b