Parser: Vary the preprocess-hash cache by extension tags available
This may not be an issue under normal circumstances, as usually
non-existent extension tags won't be used before the extension is
installed.

But it can break the wiki when upgrading from 1.39 to 1.43, since
the support of parsing the img tag as a native HTML tag is removed
from the parser between these two LTS versions.

With the preprocess-hash cache from 1.39, it's not parsed as
an extension tag, and it's difficult for server admin to clear
the cache if not using APCu as the WAN Cache.

Bug: T411459
Change-Id: Ice0a93995535f285a4a8dd1497dfbd6da2c2e28d
(cherry picked from commit 90b71e6e72ffaa54754af8fb3fd12dcecd3d9487)