findBadBlobs: Fix the --scan-to option
Currently, if --scan-to is used, the $limit can't be set which means the
total will be the default. That means $batchSize gets reduced to 0 or 1,
leading to empty query in the second round.
For example, see:
https://phabricator.wikimedia.org/P86355

By simply changing the total to INF in case of --scan-to, you make sure
the min() gets INF all the time, so the batch-size gets picked. That
also simplifies the code a bit.

Bug: T351953
Change-Id: Ib34f44364a74e903bc28f92de24983c5e4f07662