Followup I81a2c4de77: Verify stats label values are not empty
Set up a fallback value for stats label. The check that was there
did not properly verify that the value is not an empty string.
This patch adds a better check to protect against emptiness and
sets up a fallback value for latency, just in case, so it, too,
isn't null by some edge case.

Bug: T411585
Change-Id: Ie40b9e59a472affea78188bb19a083c4b432436f