UserGroupsSpecialPage: Restore invisible labels on expiry fields
While the meaning of the dropdown fields should be obvious,
it doesn't hurt to add a slightly redundant label for screen readers,
and it avoids warnings from accessibility testing software.

The label's association with the field unfortunately does not work
correctly, due to the way that the 'selectorother' combines two inputs.
This will be fixed in separate patches to OOUI and MW widgets (T411676).

This partially reverts commit 73bb8177d8e4726d1c3268a726e6b2bb5c7353a4.

Bug: T410845
Change-Id: I3fc0b08623d5b260c407ab4d8eab68bda49ce2ed