SelectWithInputWidget: Improve 'aria-labelledby' attributes
When a OOUI widget is associated with a label, the default behavior is
to set 'aria-labelledby' on the root node (this.$element). For this
widget doing that is not useful, as the root node is not focusable,
and instead contains two separate widgets, neither of which receives
a useful label.

Override setLabelledBy() to set the label on each sub-widget instead.

This requires a OOUI patch to actually fix the problem: in I24735d43df
we also override setLabelledBy() on any TabIndexedElement (otherwise
the label on TextInputWidget is similarly set on its root node, which
is not useful).

Bug: T411676
Change-Id: I9b0901f1c337f0c714062006f01f2e1ca77310b0