SelectWithInputWidget: Avoid FOUC when initial value is not "other"
The PHP widget always showed by the text input and the dropdown input,
but when using the select-or-other mode, the JS widget only shows the
text input when the value of the dropdown input is "other". As a
result, when the initial value is not "other", the text input would
appear during the loading of the page and then disappear once the JS
kicked in.

Replicate the JS behavior on the PHP side, using CSS so that it only
triggers if JS is about to be loaded (so that no-JavaScript users can
still use both inputs). This avoids the visual change once JS loads.

Update some styles for Special:UserRights that implemented the same
thing imperfectly, causing T411261. The previous code caused the
fields to be stacked vertically by accident; make this explicit.

Bug: T411261
Change-Id: I7d31daf2ec2eeff39acbc06019caa9df7ef2e710