Fix option handling for content getting postprocessed (take 2)
There was a mix-up on the options passed to postprocess and the
corresponding callback handling.

Right now it probably wasn't changing anything as the postproc option
handling is quite wide anyway, but it could have been problematic down
the line.

Ensure that the postprocessing options are properly associated with
the cloned ParserOutput when using `allowClone = true`.

As it was, it was preventing getting the correct cache key in
postprocess for a follow-up patch that adds the postprocessing cache key
in the limit report.

Bug: T410919
Change-Id: I21d4de62a9f19b3b055b7d122eceddea395ffddd