Adjust styling of confirmation emails
Unfortunately, GMail does not support svg files, and the wordmark and
tagline are only available as such. Thus, dropping them and then
simplifying the related code.

Also, <style> tags in the body are not supported in GMail, while <style>
tags in the header seem to have better support overall.
(see https://www.campaignmonitor.com/css/style-element/style-in-body/)

This always prefixes the id to overcome specificity issues in GMail,
particularly around the font-color of the button.

In general, GMail is hoped to be the most restrictive email client among
the relevant ones.

Bug: T411526
Change-Id: If4b941d6aaa579e99c832a3c25721e793bd05f84