Followup Ie40b9e59a4: Fortify unified metrics method
The original method intended to provide base setup with the
ability to override the values, but the way it was written can
also unintentionally allow for unregistered/unexpected new
labels, which should not happen, and may create errors.

There is also an issue where some edge cases seem to not fill
certain label values correctly. The new metrics should record
the "hit" on the API module, so we don't want to *ignore* that
error, but we cannot let it pass as-is because it creates
warnings. We should make sure we log these with some clearly
noted "error" state in the value so we can make sure we track
things (in grafana) and see how many (and why) these types of
API errors happen.

This commit fixes those two things:
- Makes sure labels are set with expected labels (and any
  unrecognized label is skipped)
- Record any edge case where the value is empty as an error
  so we can track those.

Bug: T411793
Change-Id: I6325499e07386c9f968b15c1cb6fa9c4f6d408dd
(cherry picked from commit 5061777a736d37e0a0f9b40359d7bb1c1f0601cb)