Api: Expand local links in html error format
Do not expect that the html error message from an api call is shown on
wiki, it could be shown in any client, like apps, where relative links
not working, use absolute links.
Or on another wiki when doing cross-site requests.

The error message does not include any styling for the error message,
but even unstyled the absolute link could be more helpful and not wrong.

Bug: T404516
Change-Id: I9368c631bfcbe2d8effe1e08c4edfe7dc86ec50f