SpecialRenameUser: Include cancel button in warning response
SpecialRenameUser may show warnings, a checkbox for ignoring the
warnings and a submit button for submission.

However, If the user decides they'd heed the warnings and abort
the rename process, there's no cancel button. Currently they've
to either use browser back button or close the tab/window completely.

This patch adds a cancel button when warnings are shown.

Change-Id: I71a3fe88bf1ef6e76790601f5e71e02590f8c4bd