Fix incomplete code path in UnregisteredLocalFile
I don't know for sure if this is the reason for the issue we see
in T412366. But I found this incompete code path while looking for
possible issues in the related code. getSizeAndMetadataWithFallback
can actually return false. All following code would break because
of that because it expects an array.

Bug: T412366
Change-Id: I16f90e8764298629eb75b88a7d4d6dc44803f2b5