DatabaseUpdater: Skip output about already applied updates
Instead of printing hundreds of lines like this:
  ...have cl_target_id field in categorylinks table.
  ...collation table already exists.
  ...module_deps doesn't exist.

Print a single line like this:
  Skipped 186 updates that were already applied.

There is still a bit of noise about some non-standard updates
that don't fit into the skipping system, but it's so much
more readable.

Most of the changes are just adding a wrapper for output() that is
only used for printing these log notices, but some code changes for
logged maintenance scripts were needed as well.

I only made changes to the output where I was pretty sure that the
update can't be applied because it was already applied, or because
some other update made it irrelevant (e.g. the table to be updated
was dropped). Some warnings about updates that don't apply are still
printed.

Change-Id: I9cae5a917fa60484ee02c37acd31417673b0157c