Rdbms: Deprecate unused IDatabase::lockIsFree footgun
It was never used after a decade in Rdbms, with two exceptions both
recent and both with better alternatives that may have arisen were
it not for this method existing. Rather than complicate the docs or
the API by other means to mitigate this, we can improve the API by
deprecating it altogether.

* DatabaseBlockStore:
  Added in I293ec64f6a (ec5cd6e2d3), fixed in I4a81f83d5a.

  Prefer a real non-blocking lock instead of lockIsFree+lock,
  which takes multiple roundtrips and is prone to a race that
  would still perform a blocking timeout.

* RecentChangesUpdateJobTest:
  Added in Id32780c00b (fc5798668d), refactored in I4313c70f8.

  Prefer stricter Database::sessionLocksPending check,
  or verify that nested jobs actually work.

Change-Id: Ie7e8e4460b714891c7fbaaafc8479d56723bb79e