Display extra form fields dynamically on Special:Log
Some log types offer additional filtering options when viewing
Special:Log. However, the relevant form fields would only be
shown after submitting the form with the selected log type (or
when visiting Special:Log/foo directly). Now they will appear
immediately when selecting the log type in the dropdown.

This is done by including all of the extra fields in the HTML
regardless of the currently selected log type, and hiding them
client-side using 'hide-if' rules.

Normally all 'hide-if' fields are shown to no-JS users, but since
there are so many of them in this form, use the new 'hide-if-nojs'
option instead to hide those that depend on the selected type.
They will be shown only after submitting the form.

Depends-On: I3a851f0e624d5f59c274c68fcb809f42e62c5ef3
Bug: T320871
Bug: T416084
Change-Id: Ie3112a4b83532f500851ec17e51b35b2f9907c4a