i18n: condense undo-summary-multiple message
The overly verbose message was brought up recently in [1]. The other
undo-summary messages say "revision" instead of "revision number", and
it's unnecessary to repeat the word 3 times. The number is linked, so
it's clear what it implies. Also, avoid use of brackets since the edit
summary is already wrapped in brackets and nested brackets are awkward.

[1]: https://en.wikipedia.org/wiki/Wikipedia:Village_pump_(technical)#Default_edit_summary_when_revering_multiple_edits

Change-Id: Ic95c4678e44be303a0d5039b04bd3717b9ada3a7