fix(ConfirmEmail): Save settings after generating tokens
Follow-up to I2f5b263b5a52ca74d2bf079f1f219c8f56103610.

Previously, User::sendConfirmationMail() was also saving settings,
which was important to persist the token. In the new version
of the code, this does not happen, resulting in invalid confirmation
emails.

This also adds a regression test to ConfirmEmailSenderTest.

Bug: T413435
Change-Id: Ie0bf35501768db8c63d09ad14333f8ca302818b4