SiteConfiguration: Optimize processSetting for default-only case
SiteConfiguration::processSetting is a hot path, consuming ~0.6% of
index.php and 1.5% of load.php wall-time in WMF production.

About one-third of WMF settings define only a 'default' value with no
overrides. This adds a short-circuit for those cases, returning the
default immediately to bypass expensive tag iteration. Even though it
only affects a minority of cases, the standard tag iteration is so
expensive that skipping it is a performance win overall.

Benchmarking getAll for enwiki on a deployment host shows a 27%
reduction in execution time:
* Before: 1.26 ms
* After:  0.92 ms

Change-Id: I906d5b3f2fc45d8e3e764409e7b2359d73fc8878